remove old widget
